+++ internal/openapi/openapi.go
@@ -1,0 +1,19 @@
+package openapi
+
+import (
+	"github.com/getkin/kin-openapi/openapi3"
+	"github.com/tufin/oasdiff/load"
+)
+
+func NewSpecInfo(path string) (*load.SpecInfo, error) {
+	loader := openapi3.NewLoader()
+	loader.IsExternalRefsAllowed = true
+	openapi3.CircularReferenceCounter = 10
+
+	spec, err := load.LoadSpecInfo(loader, load.NewSource(path))
+	if err != nil {
+		return nil, err
+	}
+
+	return spec, nil
+}
